LightLevel changes.
Removed more than half of the obtacles.
Added two new checkpoints.
Removed the dark spacing from the lights.
Added new lights to make sure that there would always be lights with
this spacing.
Increased intensity of lights.
Slowed down the running speed of Emma by 10%
Fixed her animation to match the new speed.
Added anti-aliasing to the camera.

+++ SmartButler/Assets/Scripts/Lights.cs
@@ -5,7 +5,7 @@ public class Lights : MonoBehaviour {
     private Transform[] _lightTransforms;
     private Light[] _lights;
     public int lightSpacing = 250; //spacing between light | keep the spacing low or white laggy light appear
-    public int lightIntensity = 5; //intensity of light
+    public float lightIntensity = 5; //intensity of light
 
     // Use this for initialization
     void Start(){
